+++ iss/src/rv32a_cpu.cpp
@@ -83,7 +83,7 @@ void rv32a_cpu::lrw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
@@ -111,11 +111,11 @@ void rv32a_cpu::scw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         if (rsvd_mem.active && access_addr >= rsvd_mem.start_addr && (access_addr + 3) <= rsvd_mem.end_addr)
         {
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2], MEM_WR_ACCESS_WORD, access_fault);
+            write_mem(access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2], MEM_WR_ACCESS_WORD, access_fault);
             state.hart[curr_hart].x[d->rd] = 0;
         }
         else
@@ -140,14 +140,14 @@ void rv32a_cpu::amoswapw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
         if (!access_fault)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2], MEM_WR_ACCESS_WORD, access_fault);
+            write_mem((uint32_t)access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2], MEM_WR_ACCESS_WORD, access_fault);
         }
     }
 
@@ -165,14 +165,14 @@ void rv32a_cpu::amoaddw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
         if (!access_fault)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2] + rd_val, MEM_WR_ACCESS_WORD, access_fault);
+            write_mem(access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2] + rd_val, MEM_WR_ACCESS_WORD, access_fault);
         }
     }
 
@@ -190,14 +190,14 @@ void rv32a_cpu::amoxorw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
         if (!access_fault)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2] ^ rd_val, MEM_WR_ACCESS_WORD, access_fault);
+            write_mem(access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2] ^ rd_val, MEM_WR_ACCESS_WORD, access_fault);
         }
     }
 
@@ -215,14 +215,14 @@ void rv32a_cpu::amoandw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
         if (!access_fault)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2] & rd_val, MEM_WR_ACCESS_WORD, access_fault);
+            write_mem(access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2] & rd_val, MEM_WR_ACCESS_WORD, access_fault);
         }
     }
 
@@ -240,14 +240,14 @@ void rv32a_cpu::amoorw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
         if (!access_fault)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
-            write_mem(access_addr, state.hart[curr_hart].x[d->rs2] | rd_val, MEM_WR_ACCESS_WORD, access_fault);
+            write_mem(access_addr, (uint32_t)state.hart[curr_hart].x[d->rs2] | rd_val, MEM_WR_ACCESS_WORD, access_fault);
         }
     }
 
@@ -265,7 +265,7 @@ void rv32a_cpu::amominw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
@@ -273,7 +273,7 @@ void rv32a_cpu::amominw(const p_rv32i_decode_t d)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
 
-            uint32_t wr_val = ((int32_t)rd_val < (int32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : state.hart[curr_hart].x[d->rs2];
+            uint32_t wr_val = ((int32_t)rd_val < (int32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : (uint32_t)state.hart[curr_hart].x[d->rs2];
 
             write_mem(access_addr, wr_val, MEM_WR_ACCESS_WORD, access_fault);
         }
@@ -293,7 +293,7 @@ void rv32a_cpu::amomaxw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
@@ -301,7 +301,7 @@ void rv32a_cpu::amomaxw(const p_rv32i_decode_t d)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
 
-            uint32_t wr_val = ((int32_t)rd_val > (int32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : state.hart[curr_hart].x[d->rs2];
+            uint32_t wr_val = ((int32_t)rd_val > (int32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : (uint32_t)state.hart[curr_hart].x[d->rs2];
 
             write_mem(access_addr, wr_val, MEM_WR_ACCESS_WORD, access_fault);
         }
@@ -321,7 +321,7 @@ void rv32a_cpu::amominuw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
@@ -329,7 +329,7 @@ void rv32a_cpu::amominuw(const p_rv32i_decode_t d)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
 
-            uint32_t wr_val = (rd_val < state.hart[curr_hart].x[d->rs2]) ? rd_val : state.hart[curr_hart].x[d->rs2];
+            uint32_t wr_val = (rd_val < (uint32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : (uint32_t)state.hart[curr_hart].x[d->rs2];
 
             write_mem(access_addr, wr_val, MEM_WR_ACCESS_WORD, access_fault);
         }
@@ -349,7 +349,7 @@ void rv32a_cpu::amomaxuw(const p_rv32i_decode_t d)
 
     if (!disassemble)
     {
-        access_addr = state.hart[curr_hart].x[d->rs1];
+        access_addr = (uint32_t)state.hart[curr_hart].x[d->rs1];
 
         uint32_t rd_val = read_mem(access_addr, MEM_RD_ACCESS_WORD, access_fault);
 
@@ -357,7 +357,7 @@ void rv32a_cpu::amomaxuw(const p_rv32i_decode_t d)
         {
             state.hart[curr_hart].x[d->rd] = rd_val;
 
-            uint32_t wr_val = (rd_val > state.hart[curr_hart].x[d->rs2]) ? rd_val : state.hart[curr_hart].x[d->rs2];
+            uint32_t wr_val = (rd_val > (uint32_t)state.hart[curr_hart].x[d->rs2]) ? rd_val : (uint32_t)state.hart[curr_hart].x[d->rs2];
 
             write_mem(access_addr, wr_val, MEM_WR_ACCESS_WORD, access_fault);
         }
